Added log list

+++ app/src/main/java/com/gsrajeni/appscheduler/ui/navigation/route/AppRoute.kt
@@ -12,7 +12,7 @@ sealed class AppRoute {
     object AddScheduleRoute : AppRoute()
 
     @Serializable
-    data class EditScheduleRoute(val scheduleId: Int? = null) : AppRoute()
+    data class EditScheduleRoute(val scheduleId: Long? = null) : AppRoute()
 
     @Serializable
     object ScheduleLogRoute : AppRoute()
